requirements.txt
PyPDF2
google.generativeai
pymongo
base64
streamlit_pdf_viewer
numpy
streamlit

